chowning dir

+++ recipes/vim.rb
@@ -14,6 +14,7 @@ script 'extract_module' do
 	 sudo update-alternatives --install /usr/bin/editor editor /usr/bin/nvim 60
 	 sudo update-alternatives --config editor
   EOH
+  not_if { ::File.exist?('/usr/bin/nvim') }
 end
